Make windows executable creation not run on main.

+++ build_exe.py
@@ -1,4 +1,5 @@
 import PyInstaller.__main__
+from subprocess import Popen
 from ev3sim import __version__
 
 # First, generate the version file to be used in generation.
@@ -17,4 +18,8 @@ PyInstaller.__main__.run(
 
 import os
 
-os.remove("dist/ev3sim/ev3sim/user_config.yaml")
+if os.path.exists("dist/ev3sim/ev3sim/user_config.yaml"):
+    os.remove("dist/ev3sim/ev3sim/user_config.yaml")
+
+process = Popen("makensis config.nsi")
+process.wait()
